chore: add explicit cast

+++ sensing/autoware_pointcloud_preprocessor/src/pointcloud_accumulator/pointcloud_accumulator_node.cpp
@@ -24,7 +24,8 @@ PointcloudAccumulatorComponent::PointcloudAccumulatorComponent(const rclcpp::Nod
   // set initial parameters
   {
     accumulation_time_sec_ = declare_parameter<double>("accumulation_time_sec");
-    pointcloud_buffer_.set_capacity(declare_parameter<int64_t>("pointcloud_buffer_size"));
+    pointcloud_buffer_.set_capacity(
+      static_cast<size_t>(declare_parameter<int64_t>("pointcloud_buffer_size")));
   }
 
   using std::placeholders::_1;
